Rename package service -> domain
