+++ cleanair/src/test/java/com/commento/cleanair/domain/SeoulApiServiceTest.java
@@ -1,4 +1,4 @@
-package com.commento.cleanair.service;
+package com.commento.cleanair.domain;
 
 import com.commento.cleanair.dto.AirQualityDto;
 import com.commento.cleanair.utils.utilenum.AirQualityGu;
@@ -7,8 +7,6 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import java.util.List;
 
 
 @SpringBootTest
